# specs/001-body-recomp-goals/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Body Recomposition Tracking API
  version: 1.0.0
  description: |
    RESTful API for tracking body recomposition goals including cutting and bulking phases.
    
    ## Features
    - User registration and authentication with JWT tokens
    - Multiple body fat calculation methods (Navy, 3-Site, 7-Site Skinfold)
    - Goal creation and tracking (cutting/bulking)
    - Weekly progress logging and visualization
    - Personalized training and diet plans
    
    ## Authentication
    Most endpoints require JWT bearer token authentication. Obtain tokens via `/api/v1/auth/login`.
    
  contact:
    name: Body Recomp API Support
    email: support@bodyrecomp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.bodyrecomp.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Register a new user
      description: Create a new user account with initial profile information
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              cutting_user:
                summary: User planning to cut
                value:
                  email: john.doe@example.com
                  password: SecurePass123!
                  full_name: John Doe
                  date_of_birth: "1990-05-15"
                  gender: male
                  height_cm: 175.5
                  preferred_calculation_method: navy
                  activity_level: moderately_active
              bulking_user:
                summary: User planning to bulk
                value:
                  email: jane.smith@example.com
                  password: StrongPass456!
                  full_name: Jane Smith
                  date_of_birth: "1995-08-22"
                  gender: female
                  height_cm: 165.0
                  preferred_calculation_method: 3_site
                  activity_level: very_active
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                email: john.doe@example.com
                full_name: John Doe
                date_of_birth: "1990-05-15"
                gender: male
                height_cm: 175.5
                preferred_calculation_method: navy
                activity_level: moderately_active
                created_at: "2025-10-23T10:30:00Z"
                updated_at: "2025-10-23T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    patch:
      tags:
        - Users
      summary: Update current user profile
      description: Update the authenticated user's profile information
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            examples:
              update_activity:
                summary: Update activity level
                value:
                  activity_level: very_active
              update_height:
                summary: Update height
                value:
                  height_cm: 176.0
              update_multiple:
                summary: Update multiple fields
                value:
                  full_name: John David Doe
                  activity_level: extremely_active
                  preferred_calculation_method: 7_site
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from /api/v1/auth/login
  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
        - full_name
        - date_of_birth
        - gender
        - height_cm
        - preferred_calculation_method
        - activity_level
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address (used for login)
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: User password (minimum 8 characters)
          example: SecurePass123!
        full_name:
          type: string
          maxLength: 255
          minLength: 1
          description: User's full name
          example: John Doe
        date_of_birth:
          type: string
          format: date
          description: User's date of birth (age must be 13-120 years)
          example: "1990-05-15"
        gender:
          $ref: '#/components/schemas/Gender'
        height_cm:
          type: number
          format: float
          minimum: 120
          maximum: 250
          description: Height in centimeters
          example: 175.5
        preferred_calculation_method:
          $ref: '#/components/schemas/CalculationMethod'
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 255
          minLength: 1
          example: John David Doe
        height_cm:
          type: number
          format: float
          minimum: 120
          maximum: 250
          example: 176.0
        preferred_calculation_method:
          $ref: '#/components/schemas/CalculationMethod'
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
      description: All fields are optional. Only provided fields will be updated.
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        full_name:
          type: string
          example: John Doe
        date_of_birth:
          type: string
          format: date
          example: "1990-05-15"
        gender:
          $ref: '#/components/schemas/Gender'
        height_cm:
          type: number
          format: float
          example: 175.5
        preferred_calculation_method:
          $ref: '#/components/schemas/CalculationMethod'
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-10-23T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last profile update timestamp
          example: "2025-10-23T10:30:00Z"
...
    Gender:
      type: string
      enum:
        - male
        - female
      description: User's biological gender (used for body fat calculations)
      example: male
    CalculationMethod:
      type: string
      enum:
        - navy
        - 3_site
        - 7_site
      description: |
        Body fat calculation method:
        - navy: US Navy Circumference Method (requires measuring tape only)
        - 3_site: Jackson-Pollock 3-site skinfold (requires calipers)
        - 7_site: Jackson-Pollock 7-site skinfold (most accurate, requires calipers)
      example: navy
    ActivityLevel:
      type: string
      enum:
        - sedentary
        - lightly_active
        - moderately_active
        - very_active
        - extremely_active
      description: |
        User's activity level for TDEE calculation:
        - sedentary: Little to no exercise
        - lightly_active: Light exercise 1-3 days/week
        - moderately_active: Moderate exercise 3-5 days/week
        - very_active: Hard exercise 6-7 days/week
        - extremely_active: Very hard exercise, physical job, or training twice per day
      example: moderately_active
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: The requested resource was not found
        error_code:
          type: string
          description: Machine-readable error code
          example: RESOURCE_NOT_FOUND
        field:
          type: string
          nullable: true
          description: Field name if validation error
          example: target_body_fat_percentage
        value:
          nullable: true
          description: Invalid value if applicable
          example: 3.5
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
                example: ["body", "height_cm"]
              msg:
                type: string
                description: Validation error message
                example: ensure this value is greater than or equal to 120
              type:
                type: string
                description: Error type
                example: value_error.number.not_ge
              ctx:
                type: object
                nullable: true
                description: Additional context about the error
  responses:
    BadRequest:
      description: Bad request - Invalid input or business logic violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unsafe_target:
              summary: Unsafe body fat target
              value:
                detail: "Target body fat percentage is below safe threshold for males (8%)"
                error_code: UNSAFE_BODY_FAT_TARGET
                field: target_body_fat_percentage
                value: 5.0
            missing_measurements:
              summary: Missing required measurements
              value:
                detail: "Navy method requires waist_cm and neck_cm measurements"
                error_code: MISSING_REQUIRED_MEASUREMENTS
                field: null
                value: null
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                detail: "Authentication token is required"
                error_code: AUTHENTICATION_REQUIRED
            invalid_token:
              summary: Invalid or expired token
              value:
                detail: "Authentication token is invalid or expired"
                error_code: INVALID_TOKEN
            invalid_credentials:
              summary: Invalid login credentials
              value:
                detail: "Incorrect email or password"
                error_code: INVALID_CREDENTIALS
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            detail:
              - loc: ["body", "height_cm"]
                msg: "ensure this value is greater than or equal to 120"
                type: "value_error.number.not_ge"
                ctx:
                  limit_value: 120
              - loc: ["body", "email"]
                msg: "value is not a valid email address"
                type: "value_error.email"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "An unexpected error occurred. Please try again later."
            error_code: INTERNAL_SERVER_ERROR